old compiler fix

+++ crypto/params.c
@@ -760,7 +760,7 @@ int OSSL_PARAM_get_uint64(const OSSL_PARAM *p, uint64_t *val)
                      * point values.
                      */
                     && d < (double)(UINT64_MAX - 65535) + 65536.0
-                    && d == (uint64_t)d) {
+                    && d == (__int64)d) {
                 *val = (uint64_t)d;
                 return 1;
             }
@@ -822,7 +822,7 @@ int OSSL_PARAM_set_uint64(OSSL_PARAM *p, uint64_t val)
         switch (p->data_size) {
         case sizeof(double):
             if ((val >> real_shift()) == 0) {
-                *(double *)p->data = (double)val;
+                *(double *)p->data = (double)(__int64)val;
                 return 1;
             }
             break;
@@ -971,7 +971,7 @@ int OSSL_PARAM_get_double(const OSSL_PARAM *p, double *val)
         case sizeof(uint64_t):
             u64 = *(const uint64_t *)p->data;
             if ((u64 >> real_shift()) == 0) {
-                *val = (double)u64;
+                *val = (double)(__int64)u64;
                 return 1;
             }
             break;
@@ -1010,7 +1010,7 @@ int OSSL_PARAM_set_double(OSSL_PARAM *p, double val)
             return 1;
         }
     } else if (p->data_type == OSSL_PARAM_UNSIGNED_INTEGER
-               && val == (uint64_t)val) {
+               && val == (__int64)val) {
         p->return_size = sizeof(double);
         if (p->data == NULL)
             return 1;
